chore: remove cpu crate
