--- crates/cpu/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "cpu"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-aarch64-cpu = "9.4.0"
