+++ crates/device/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-cpu = { version = "0.1.0", path = "../cpu" }
 paste = "1.0.14"
 small-std = { version = "0.1.0", path = "../small-std" }
 tock-registers = "0.9.0"
