+++ crates/kernel/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-cpu = { version = "0.1.0", path = "../cpu" }
 small-std = { version = "0.1.0", path = "../small-std" }
 panic-wait = { version = "0.1.0", path = "../panic-wait" }
 device = { version = "0.1.0", path = "../device" }
