+++ crates/panic-wait/Cargo.toml
@@ -6,5 +6,4 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-cpu = { version = "0.1.0", path = "../cpu" }
 small-std = { version = "0.1.0", path = "../small-std" }
