+++ crates/panic-wait/src/lib.rs
@@ -3,12 +3,13 @@
 #![feature(panic_info_message)]
 #![no_std]
 
+mod utils;
 use core::panic::PanicInfo;
 use small_std::println;
 
 #[panic_handler]
 fn panic(info: &PanicInfo) -> ! {
-    panic_prevent_reenter();
+    utils::panic_prevent_reenter();
 
     let (location, line, column) = match info.location() {
         Some(location) => (location.file(), location.line(), location.column()),
@@ -25,20 +26,5 @@ fn panic(info: &PanicInfo) -> ! {
         info.message().unwrap()
     );
 
-    cpu::wait_forever();
-}
-
-/// Stop immediately if called a second time.
-fn panic_prevent_reenter() {
-    // Using atoms can save us from needing `unsafe` here.
-    use core::sync::atomic::{AtomicBool, Ordering};
-
-    static PANIC_IN_PROGRESS: AtomicBool = AtomicBool::new(false);
-
-    if !PANIC_IN_PROGRESS.load(Ordering::Relaxed) {
-        PANIC_IN_PROGRESS.store(true, Ordering::Relaxed);
-        return;
-    }
-
-    cpu::wait_forever();
+    utils::wait_forever();
 }
